Fix: createpage memoization
